rename `exceptions` to `errors`
