+++ python/packages/ai/teams/ai/__init__.py
@@ -4,7 +4,8 @@ Licensed under the MIT License.
 """
 
 from .ai import *
+from .ai_error import *
+from .ai_options import *
 from .history_options import *
-from .options import *
 from .prompts import *
 from .state import *
