+++ python/packages/ai/teams/ai/ai_error.py
@@ -4,6 +4,6 @@ Licensed under the MIT License.
 """
 
 
-class ApplicationException(Exception):
+class AIError(Exception):
     def __init__(self, message: str) -> None:
         super().__init__(message)
