--- python/packages/ai/teams/ai/exceptions/__init__.py
@@ -1,7 +1,0 @@
-"""
-Copyright (c) Microsoft Corporation. All rights reserved.
-Licensed under the MIT License.
-"""
-
-from .ai import *
-from .prompt_manager import *
