+++ python/packages/ai/teams/ai/prompts/__init__.py
@@ -4,3 +4,4 @@ Licensed under the MIT License.
 """
 
 from .prompt_manager import *
+from .prompt_manager_error import *
