+++ python/packages/ai/teams/ai/prompts/prompt_manager_error.py
@@ -4,6 +4,6 @@ Licensed under the MIT License.
 """
 
 
-class AIException(Exception):
+class PromptManagerError(Exception):
     def __init__(self, message: str) -> None:
         super().__init__(message)
